Employee Management System
In this project there are multiple admin which can login into the
system. Admin can manage the employees and department.Crud operations
can be performed on both employee and department.This project is
developed using Spring MVC  and Framework and using Hibernate as an ORM
tool. MySQL is used for database.I have also used Maven for project
management.